+++ src/executable/command.rs
@@ -1,0 +1,4 @@
+use semver::{Version, VersionReq};
+
+use super::Executable;
+use crate::error::*;
