+++ src/source.rs
@@ -17,16 +17,17 @@ pub enum CrateType {
 }
 
 #[derive(Hash, Clone)]
+/// Information about CUDA crate.
 pub struct Crate {
     path: PathBuf,
-
     crate_type: CrateType,
     output_file_prefix: String,
     deps_file_prefix: String,
 }
 
 impl Crate {
-    pub fn analyze<P: AsRef<Path>>(path: P) -> Result<Self> {
+    /// Try to locate a crate at the `path` and collect needed information.
+    pub fn analyse<P: AsRef<Path>>(path: P) -> Result<Self> {
         let path = env::current_dir()?.join(&path);
 
         match fs::metadata(path.join("Cargo.toml")) {
@@ -79,18 +80,22 @@ impl Crate {
         })
     }
 
+    /// Returns PTX assmbly filename prefix.
     pub fn get_output_file_prefix(&self) -> &str {
         &self.output_file_prefix
     }
 
+    /// Returns deps file filename prefix.
     pub fn get_deps_file_prefix(&self) -> &str {
         &self.deps_file_prefix
     }
 
+    /// Returns crate root path.
     pub fn get_path(&self) -> &Path {
         &self.path.as_path()
     }
 
+    /// Returns temporary crate build location.
     pub fn get_output_path(&self) -> Result<PathBuf> {
         let mut path = env::temp_dir().join("ptx-builder-0.5");
 
@@ -107,4 +112,61 @@ impl Crate {
 
         hasher.finish()
     }
+}
+
+#[test]
+fn should_find_crate_names() {
+    let source = Crate::analyse("tests/fixtures/sample-crate").unwrap();
+
+    assert_eq!(source.get_output_file_prefix(), "sample_ptx_crate");
+    assert_eq!(source.get_deps_file_prefix(), "libsample_ptx_crate");
+}
+
+#[test]
+fn should_find_app_crate_names() {
+    let source = Crate::analyse("tests/fixtures/app-crate").unwrap();
+
+    assert_eq!(source.get_output_file_prefix(), "sample_app_ptx_crate");
+    assert_eq!(source.get_deps_file_prefix(), "sample_app_ptx_crate");
+}
+
+#[test]
+fn should_check_existence_of_crate_path() {
+    let result = Crate::analyse("tests/fixtures/non-existing-crate");
+
+    match result {
+        Err(Error(ErrorKind::InvalidCratePath(path), _)) => {
+            assert!(path.ends_with("tests/fixtures/non-existing-crate"));
+        }
+
+        Ok(_) => unreachable!("it should fail"),
+        Err(_) => unreachable!("it should fail with proper error"),
+    }
+}
+
+#[test]
+fn should_check_validity_of_crate_path() {
+    let result = Crate::analyse("tests/builder.rs");
+
+    match result {
+        Err(Error(ErrorKind::InvalidCratePath(path), _)) => {
+            assert!(path.ends_with("tests/builder.rs"));
+        }
+
+        Ok(_) => unreachable!("it should fail"),
+        Err(_) => unreachable!("it should fail with proper error"),
+    }
+}
+
+#[test]
+fn should_provide_output_path() {
+    let source_crate = Crate::analyse("tests/fixtures/sample-crate").unwrap();
+
+    assert!(
+        source_crate.get_output_path().unwrap().starts_with(
+            env::temp_dir()
+                .join("ptx-builder-0.5")
+                .join("sample_ptx_crate")
+        )
+    );
 }
